custom hook to fetch the user
