+++ src/hook/useUser.ts
@@ -1,0 +1,26 @@
+import { useEffect, useState } from 'react';
+import { getUserData } from '../services/user.api';
+import { IUser } from '../types/types';
+
+const useUser = () => {
+  const [userData, setUserData] = useState<IUser | null>(null);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const result = await getUserData();
+        setUserData(result);
+      } catch (error) {
+        console.error(error);
+      } finally {
+        setLoading(false);
+      }
+    };
+    fetchData();
+  }, []);
+
+  return { userData, loading };
+};
+
+export default useUser;
